refactor: rename `internal/fileutil` into `internal/fsutil`
