+++ internal/fsext/fileutil.go
@@ -1,4 +1,4 @@
-package fileutil
+package fsext
 
 import (
 	"fmt"
